eunice-hong/src/test/kotlin/com/eunicehong/Day03Test.kt
package com.eunicehong

import org.junit.jupiter.api.Assertions.assertEquals
import org.junit.jupiter.api.Test

/**
 * Check if the solutions for [Day03] is correct.
 */
internal class Day03Test : DailyTest(3) {
    private val testDay03: Day03 = Day03()

    @Test
    override fun puzzleSolution1() {
        val expected = "162813399"
        val actual = testDay03.solution1(input)
        assertEquals(expected, actual, "🎅 Solution for Puzzle 1 Day $dayString has not been implemented yet.")
        println("🌲 Day $dayString Puzzle 1 completed! $actual")
    }

    @Test
    override fun puzzleSolution2() {
        val expected = "53783319"
        val actual = testDay03.solution2(input)
        assertEquals(expected, actual, "🎅 Solution for Puzzle 2 Day $dayString has not been implemented yet.")
        println("🌲 Day $dayString Puzzle 2 completed! $actual")
    }
}
